Fixed Linters Errors
